selectSpace trigger and listenTo are functioning and passing thru space id
